% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character_allStats.R
\name{allStats}
\alias{allStats}
\title{allStats}
\source{
A minimal web application that extracts statistical results from single documents with [get.stats()] is hosted at: \href{www.get-stats.app}{https://www.get-stats.app/}
}
\usage{
allStats(x)
}
\arguments{
\item{x}{text string}
}
\value{
Vector with sticked results.
}
\description{
Extracts any statistical results from text string with some uniformizations as vector. This function is implemented in [get.stas()].
}
\examples{
x<-c("The mean difference of scale A was significant (beta=12.9, t(18)=2.5, p<.05)",
"The ANOVA yielded significant results on factor A (F(2,18)=6, p<.05, eta(g)2<-.22).",
"The correlation of x and y was r=.37.")
allStats(x)
}
\references{
Böschen (2021). "Evaluation of JATSdecoder as an automated text extraction tool for statistical results in scientific reports.” \emph{Scientific Reports.} doi: \href{https://www.nature.com/articles/s41598-021-98782-3}{10.1038/s41598-021-98782-3}.
}
