% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JATS_get.abstract.R
\name{get.abstract}
\alias{get.abstract}
\title{get.abstract}
\usage{
get.abstract(x, remove.title = TRUE, letter.convert = TRUE, cermine = FALSE)
}
\arguments{
\item{x}{a NISO-JATS coded XML file or text}

\item{remove.title}{Logical. If TRUE removes section titles in abstract}

\item{letter.convert}{Logical. If TRUE converts hex and html coded characters to unicode}

\item{cermine}{Logical. If TRUE and letter.convert=TRUE performs CERMINE specific text correction}
}
\description{
Extract abstract tag from NISO-JATS coded XML file or text as vector of abstracts
}
\examples{
x<-"Some text <abstract>Some abstract</abstract> some text"
get.abstract(x)
x<-"Some text <abstract>Some abstract</abstract> TEXT <abstract>
Some other abstract</abstract> Some text "
get.abstract(x)
}
