% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JATS_get.aff.R
\name{get.aff}
\alias{get.aff}
\title{get.aff}
\usage{
get.aff(x, remove.html = FALSE, letter.convert = TRUE)
}
\arguments{
\item{x}{a NISO-JATS coded XML file or text}

\item{remove.html}{Logical. If TRUE removes all html tags}

\item{letter.convert}{Logical. If TRUE converts hex and html coded characters to unicode}
}
\description{
Extract affiliation tag/s from NISO-JATS coded XML file or text as vector of affiliations
}
\examples{
x<-"Some text <aff>Some affiliation</aff> some text"
get.aff(x)
x<-"Some text <aff>Some affiliation</aff> TEXT <aff>Some other affiliation</aff> Some text "
get.aff(x)
}
