% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character_get.alpha.error.R
\name{get.alpha.error}
\alias{get.alpha.error}
\title{get.alpha.error}
\usage{
get.alpha.error(x)
}
\arguments{
\item{x}{text to process}
}
\description{
Extract reported alpha error from text
}
\examples{
x<-c("The threshold for significance was adjusted to .05/2",
"Type 1 error rate was alpha=.05.")
get.alpha.error(x)
x<-c("We used p<.05 as level of significance.",
     "We display .95 CIs and use an adjusted alpha of .10/3.",
     "The effect was significant with p<.025.")
get.alpha.error(x)
}
