% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JATS_get.author.R
\name{get.author}
\alias{get.author}
\title{get.author}
\usage{
get.author(x, paste = "", short.names = FALSE, letter.convert = FALSE)
}
\arguments{
\item{x}{a NISO-JATS coded XML file or text}

\item{paste}{if "" author list is exported as vector with length of number of authors, else collapsed to one cell}

\item{short.names}{Logical. If TRUE fully available first names will be reduced to one letter abbreviation}

\item{letter.convert}{Logical. If TRUE converts hex and html coded characters to unicode}
}
\description{
Extract author tag/s from NISO-JATS coded XML file or text as vector of authors
}
