% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JATS_get.category.R
\name{get.category}
\alias{get.category}
\title{get.category}
\usage{
get.category(x)
}
\arguments{
\item{x}{a NISO-JATS coded XML file or text.}
}
\value{
Character vector with the extracted category name/s.
}
\description{
Extracts category tag/s from NISO-JATS coded XML file or text as vector of categories.
}
\examples{
x<-"Some text <article-categories>Some category</article-categories> some text"
get.category(x)
}
\seealso{
\code{\link[JATSdecoder]{JATSdecoder}} for simultaneous extraction of meta-tags, abstract, sectioned text and reference list.
}
