% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JATS_get.history.R
\name{get.history}
\alias{get.history}
\title{get.history}
\usage{
get.history(x, remove.na = FALSE)
}
\arguments{
\item{x}{a NISO-JATS coded XML file or text.}

\item{remove.na}{Logical. If TRUE hides non available date stamps.}
}
\value{
Character vector with the extracted dates of publishing history.
}
\description{
Extracts available publishing history tags from NISO-JATS coded XML file or text and compute pubDate and pubyear.
}
\seealso{
\code{\link[JATSdecoder]{JATSdecoder}} for simultaneous extraction of meta-tags, abstract, sectioned text and reference list.
}
