% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JATS_get.journal.R
\name{get.journal}
\alias{get.journal}
\title{get.journal}
\usage{
get.journal(x)
}
\arguments{
\item{x}{a NISO-JATS coded XML file or text.}
}
\value{
Character string with the extracted journal name.
}
\description{
Extracts journal tag from NISO-JATS coded XML file or text.
}
\examples{
x<-"Some text <journal-title>PLoS One</journal-title> some text"
get.journal(x)
}
\seealso{
\code{\link[JATSdecoder]{JATSdecoder}} for simultaneous extraction of meta-tags, abstract, sectioned text and reference list.
}
