% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JATS_get.keywords.R
\name{get.keywords}
\alias{get.keywords}
\title{get.keywords}
\usage{
get.keywords(
  x,
  paste = "",
  letter.convert = TRUE,
  include.max = length(keyword)
)
}
\arguments{
\item{x}{a NISO-JATS coded XML file or text}

\item{paste}{if paste!="" author vector is collapsed to one cell}

\item{letter.convert}{Logical. If TRUE converts hex and html coded characters to unicode}

\item{include.max}{a maximum number of keywords to extract}
}
\description{
Extract keyword tag/s from NISO-JATS coded XML file or text as vector of keywords
}
\examples{
x<-"Some text <kwd>Keyword 1</kwd>, <kwd>Keyword 2</kwd> some text"
get.keywords(x)
get.keywords(x,paste(", "))
}
