% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character_get.multi.comparison.R
\name{get.multi.comparison}
\alias{get.multi.comparison}
\title{get.multi.comparison}
\usage{
get.multi.comparison(x)
}
\arguments{
\item{x}{text to process}
}
\description{
Extract alpha-/p-value correction method for multiple comparisons from list with 14 correction methods
}
\examples{
x<-"We used Bonferroni corrected p-values."
get.multi.comparison(x)
}
