% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character_get.n.studies.R
\name{get.n.studies}
\alias{get.n.studies}
\title{get.n.studies}
\usage{
get.n.studies(x, tolower = TRUE)
}
\arguments{
\item{x}{text string to process.}

\item{tolower}{Logical. If TRUE lowerises text and search patterns for processing.}
}
\value{
Numeric number of identified number of studies.
}
\description{
Extracts number of studies/experiments from text.
}
