% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character_get.power.R
\name{get.power}
\alias{get.power}
\title{get.power}
\usage{
get.power(x)
}
\arguments{
\item{x}{text string to process.}
}
\value{
Numeric. Identified power values.
}
\description{
Extracts a priori power and empirial power values from text.
}
\examples{
x<-"We used G*Power 3 to calculate the needed sample with 
beta error rate set to 12\% and alpha error to .05."
get.power(x)
}
