% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character_get.R.package.R
\name{get.R.package}
\alias{get.R.package}
\title{get.R.package}
\usage{
get.R.package(x, update.package.list = FALSE)
}
\arguments{
\item{x}{text to process}

\item{update.package.list}{Logical. If TRUE update of list with available packages is downloaded from CRAN with available.packages()}
}
\description{
Extract mentioned R package from text
}
\examples{
get.R.package("We used the R Software packages lme4 (and psych).")
}
