% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JATS_get.references.R
\name{get.references}
\alias{get.references}
\title{get.references}
\usage{
get.references(
  x,
  letter.convert = FALSE,
  remove.html = FALSE,
  extract = "full"
)
}
\arguments{
\item{x}{a NISO-JATS coded XML file or text}

\item{letter.convert}{Logical. If TRUE converts hex and html coded characters to unicode}

\item{remove.html}{Logical. If TRUE removes all html tags}

\item{extract}{part of refernces to extract (one of "full" or "title")}
}
\description{
Extract reference list from NISO-JATS coded XML file or text as vector of references
}
