% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_get.sentence.with.pattern.R
\name{get.sentence.with.pattern}
\alias{get.sentence.with.pattern}
\title{get.sentence.with.pattern}
\usage{
get.sentence.with.pattern(x, patterns = c(""), tolower = TRUE)
}
\arguments{
\item{x}{text to process}

\item{patterns}{search terms}

\item{tolower}{Logical. If TRUE converts search terms and text to lower case}
}
\description{
Return lines with search term patterns
}
\examples{
text<-c("This demo demonstrates how get.sentence.with.pattern works.","The is a simple 0, 1.")
get.sentence.with.pattern(text,c("Demo","example","work"))
get.sentence.with.pattern(text,c("Demo","example","work"),tolower=TRUE)
}
