% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JATS_get.subject.R
\name{get.subject}
\alias{get.subject}
\title{get.subject}
\usage{
get.subject(x, letter.convert = TRUE, paste = "")
}
\arguments{
\item{x}{a NISO-JATS coded XML file or text}

\item{letter.convert}{Logical. If TRUE converts hex and html coded characters to unicode}

\item{paste}{if "" author list is exported as vector with length of number of authors, else collapsed to one cell}
}
\description{
Extract subject tag/s from NISO-JATS coded XML file or text as vector of subjects
}
\examples{
x<-"Some text <subject>Some subject</subject> some text"
get.subject(x)
x<-"Some text <subject>Some subject</subject> TEXT <subject>Some other subject</subject> Some text "
get.subject(x)
get.subject(x,paste=", ")
}
