% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JATS_get.tables.R
\name{get.tables}
\alias{get.tables}
\title{get.tables}
\usage{
get.tables(x)
}
\arguments{
\item{x}{HTML file or text with html tags.}
}
\value{
Character vector with extracted HTML tables.
}
\description{
Extracts HTML tables as vector of tables.
}
\seealso{
\code{\link[JATSdecoder]{JATSdecoder}} for simultaneous extraction of meta-tags, abstract, sectioned text and reference list.
}
