% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character_get.test.direction.R
\name{get.test.direction}
\alias{get.test.direction}
\title{get.test.direction}
\usage{
get.test.direction(x)
}
\arguments{
\item{x}{text string to process.}
}
\value{
Character.
}
\description{
Extracts mentioned test direction/s (one sided, two sided, one and two sided) from text.
}
