% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JATS_get.title.R
\name{get.title}
\alias{get.title}
\title{get.title}
\usage{
get.title(x)
}
\arguments{
\item{x}{a NISO-JATS coded XML file or text.}
}
\value{
Character string with extracted article title.
}
\description{
Extracts article title from NISO-JATS coded XML file or text.
}
