% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JATS_get.type.R
\name{get.type}
\alias{get.type}
\title{get.type}
\usage{
get.type(x)
}
\arguments{
\item{x}{a NISO-JATS coded XML file or text.}
}
\value{
Character string with extracted article type.
}
\description{
Extracts article type from NISO-JATS coded XML file or text.
}
