% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JATS_get.vol.R
\name{get.vol}
\alias{get.vol}
\title{get.vol}
\usage{
get.vol(x)
}
\arguments{
\item{x}{a NISO-JATS XML coded file or text.}
}
\value{
Character string with extracted journal volume.
}
\description{
Extracts volume, first and last page from NISO-JATS coded XML file or text.
}
\seealso{
\code{\link[JATSdecoder]{JATSdecoder}} for simultaneous extraction of meta-tags, abstract, sectioned text and reference list.
}
