% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_grep2.R
\name{grep2}
\alias{grep2}
\title{grep2}
\usage{
grep2(pattern, x, value = TRUE, invert = FALSE, perl = FALSE)
}
\arguments{
\item{pattern}{Character string containing a regular expression (or character string for fixed = TRUE) to be matched in the given character vector. Coerced by as.character to a character string if possible. If a character vector of length 2 or more is supplied, the first element is used with a warning. Missing values are allowed except for regexpr, gregexpr and regexec.}

\item{x}{A character vector where matches are sought, or an object which can be coerced by as.character to a character vector. Long vectors are supported.}

\item{value}{Logical. if FALSE, a vector containing the (integer) indices of the matches determined by grep is returned, and if TRUE, a vector containing the matching elements themselves is returned.}

\item{invert}{Logical. If TRUE return indices or values for elements that do not match.}

\item{perl}{Logical. Should Perl-compatible regexps be used?}
}
\description{
Extension of grep(). Allows to identify and extract cells with/without multiple search patterns that are connected with AND.
}
\examples{
x<-c("ab","ac","ad","bc","bad")
grep2(c("a","b"),x)
grep2(c("a","b"),x,invert=TRUE)
grep2(c("a","b"),x,value=FALSE)
}
