% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character_has.interaction.R
\name{has.interaction}
\alias{has.interaction}
\title{has.interaction}
\usage{
has.interaction(x)
}
\arguments{
\item{x}{text to process}
}
\description{
Identify interaction/moderator/mediator effect in text
}
