% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_has.pattern.R
\name{has.pattern}
\alias{has.pattern}
\title{has.pattern}
\usage{
has.pattern(x, patterns = c(""), tolower = TRUE)
}
\arguments{
\item{x}{text to process}

\item{patterns}{search terms}

\item{tolower}{Logical. If TRUE converts search terms and text to lower case}
}
\description{
Return search term hit vector for all search patterns
}
\examples{
text<-c("This demo demonstrates how has.pattern() works.",
        "The result is a simple 0, 1 coded vector for all search patterns.")
has.pattern(text,c("Demo","example","work"))
has.pattern(text,c("Demo","example","work"),tolower=TRUE)
}
