% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_ngramaroundpattern.R
\name{ngramaroundpattern}
\alias{ngramaroundpattern}
\title{ngramaroundpattern}
\usage{
ngramaroundpattern(x, pattern, ngram = c(-3, 3), tolower = TRUE)
}
\arguments{
\item{x}{text to process}

\item{pattern}{a search string pattern to build the ngram}

\item{ngram}{a vector of length=2 that defines the number of gram on left and right side of pattern word match}

\item{tolower}{Logical. If TRUE converts text and pattern to lower case}
}
\description{
Extract an ngram word vector around a pattern word match
}
\examples{
text<-"One hundred twenty-eight students participated in our Study, 
that was administred in thirteen clinics."
ngramaroundpattern(text,pattern="study",ngram=c(-1,2))
}
