% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_preCheck.R
\name{preCheck}
\alias{preCheck}
\title{preCheck}
\usage{
preCheck(x)
}
\arguments{
\item{x}{anything}
}
\value{
The original input, error messages, empty objects or the read file content with readLines().
}
\description{
Performs prechecks and returns error messages for wrong input. If x is a textual input, x is returned unprocessed. If x is a file, x is read with readLines() and returned.
}
\keyword{internal}
