% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_strsplit2.R
\name{strsplit2}
\alias{strsplit2}
\title{strsplit2}
\usage{
strsplit2(x, split, type = "remove", perl = FALSE)
}
\arguments{
\item{x}{text to process}

\item{split}{pattern to split text at}

\item{type}{one out of "remove", "before", "after"}

\item{perl}{Logical. If TRUE uses perl expressions}
}
\description{
Extension of strsplit(). Makes it possible to split lines "before" or "after" a pattern match
}
\examples{
x<-"This is some text, where text is the split pattern of the text."
strsplit2(x,"text","after")
}
