% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.character.R
\name{study.character}
\alias{study.character}
\title{study.character}
\usage{
study.character(
  x,
  text.mode = 1,
  stats.mode = "all",
  recalculate.p = TRUE,
  alternative = "auto",
  estimateZ = FALSE,
  T2t = FALSE,
  update.package.list = FALSE,
  add.software = NULL,
  quantileDF = 0.75,
  N.max.only = FALSE,
  output = "all"
)
}
\arguments{
\item{x}{JATS coded XML file or JATSdecoder result}

\item{text.mode}{text parts to extract statistical results from (text.mode=1: abstract and full text, text.mode=2: method and result section, text.mode=3: result section only)}

\item{stats.mode}{Character. Select subset of standard stats. One of: "all", "checkable", "computable"}

\item{recalculate.p}{Logical. If TRUE recalculates p values (for 2 sided test) if possible}

\item{alternative}{Character. Select sidedness of recomputed p-values for t-, r- and Z-values. One of c("auto","undirected","directed","both"). If set to "auto" alternative will be be set to both if get.test.direction() detected one-directional hypotheses/tests. If no directional hypotheses/tests are only "undirected" recomputed p-values will be returned}

\item{estimateZ}{Logical. If TRUE detected beta-/d-value is divided by reported standard error "SE" to estimate Z-value ("Zest") for observed beta/d and recompute p-value. Note: This is only valid, if Gauss-Marcov assumptions are met and a sufficiently large sample size is used. If a Z- or t-value is detected in a report of a beta-/d-coefficient with SE, no estimation will be performed, although set to TRUE.}

\item{T2t}{Logical. If TRUE capital letter T is treated as small letter t}

\item{update.package.list}{if TRUE updates available R packages with available.packages() function}

\item{add.software}{additional software names to detect as vector}

\item{quantileDF}{quantile of (df1+1)+(df2+1) to extract for estimating sample size}

\item{N.max.only}{return only maximum of estimated sample sizes}

\item{output}{output selection of specific results c("all", "doi", "title", "year", "n.studies", "methods", "alpha.error", "power", "multi.comparison.correction", "assumptions", "OutlierRemovalInSD", "InteractionModeratorMediatorEffect", "test.direction", "sig.adjectives", "software", "Rpackage", "stats", "standardStats", "estimated.sample.size")}
}
\description{
extracts study characteristics out of a JATS coded XML file or JATSdecoder result
}
\examples{
URL <- "https://journals.plos.org/plosone/article/file?id=10.1371/journal.pone.0114876&type=manuscript"
download.file(URL,"file.xml")
JATSdecoder("file.xml")
study.character("file.xml")
study.character(JATSdecoder("file.xml"))
study.character("file.xml",out=c("stats","standardStats"))
}
