% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character_study.type.R
\name{study.type}
\alias{study.type}
\title{study.type}
\usage{
study.type(title = NULL, text = NULL)
}
\arguments{
\item{title}{articles title text}

\item{text}{main text to process}
}
\description{
function to identify type of study by list of study types
}
\examples{
study.type("We performed a randomized treatment control trail with waiting group")
}
