% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_text2sentences.R
\name{text2sentences}
\alias{text2sentences}
\title{text2sentences}
\usage{
text2sentences(x)
}
\arguments{
\item{x}{text to process}
}
\description{
Convert floating text to a vector with sentences via fine tuned regular expressions or NLP sentence tokenization
}
\examples{
x<-"Some text with result (t(18)=1.2, p<.05). This shows how text2sentences works."
text2sentences(x)
}
