% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_vectorize.text.R
\name{vectorize.text}
\alias{vectorize.text}
\title{vectorize.text}
\usage{
vectorize.text(x)
}
\arguments{
\item{x}{text to vectorize}
}
\description{
convert vector of text to a list of words within each cell. Note: punctuation will be removed
}
\examples{
text<-"One hundred twenty-eight students participated in our Study, 
that was administred in thirteen clinics."
vectorize.text(text)
}
