% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character_allStats.R
\name{allStats}
\alias{allStats}
\title{allStats}
\usage{
allStats(x)
}
\arguments{
\item{x}{text to extract statistical results from}
}
\description{
Extract statistical results from text with some uniformisation
}
\examples{
x<-c("The mean difference of scale A was significant (beta=12.9, t(18)=2.5, p<.05)",
"The ANOVA yielded significant results on 
 faktor A (F(2,18)=6, p<.05, eta(g)2<-.22)",
"the correlation of x and y was r=.37.")
allStats(x)
}
