% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est.ss.R
\name{est.ss}
\alias{est.ss}
\title{est.ss}
\usage{
est.ss(
  abstract = NULL,
  text = NULL,
  quantileDF = 0.75,
  max.only = FALSE,
  max.parts = TRUE
)
}
\arguments{
\item{abstract}{abstracts text}

\item{text}{main text to process (usually method section)}

\item{quantileDF}{quantile of (df1+1)+(df2+2) to extract}

\item{max.only}{Logical. If TRUE only the final estimate will be returned, if FALSE all sub estimates are returned as well}

\item{max.parts}{Logical. If FALSE outputs all captured sample sizes in sub inputs}
}
\description{
Function to estimate studies sample size by maximizing different conservative estimates.
Performs 4 different heuristical searches for sample size in abstract, text, stats and standardStats.
}
\details{
Sample size extraction from abstract:\cr
- Extracts N= from abstract if possible

Sample size extraction from statistical results:\cr
- Extracts "N=" in statistical result lines from get.stats() that contain p-value

Sample size extraction with result of standardStats(get.stats()):\cr
- Extracts df1 and df2 if possible and neither containing a ".", than calculates quantile of (df1+1)+(df2+2) (at least 2 group comparison assumed)
}
\examples{
 a<-"One hundred twelve students participated in our study."
 x<-"Our sample consists of three hundred twenty five undergraduate students.
     For one sub group the F-test indicates significant differences in means F(2,102)=3.21, p<.05."
 est.ss(abstract=a,text=x)
}
