% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character_get.alpha.error.R
\name{get.alpha.error}
\alias{get.alpha.error}
\title{get.alpha.error}
\usage{
get.alpha.error(x)
}
\arguments{
\item{x}{text to process}
}
\description{
Extract reported alpha error from text
}
\examples{
x<-c("The threshold for significance was adjusted to .05/2",
"Type 1 error rate was alpha=.05.")
get.alpha.error(x)
x<-"We display 95\% CIs and use an alpha error of .05/5."
get.alpha.error(x)
}
