% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character_get.assumption.R
\name{get.assumptions}
\alias{get.assumptions}
\title{get.assumptions}
\usage{
get.assumptions(x, hits_only = TRUE)
}
\arguments{
\item{x}{text string to process.}

\item{hits_only}{Logical. If TRUE returns the detected assumtions only, else a hit matrix with all potential assumptions is returned.}
}
\value{
Character. Vector with identified statistical assumption/s.
}
\description{
Extracts the mentioned statistical assumptions from a text string by a dictionary search of 22 common statistical assumptions.
}
\examples{
x<-"Sphericity assumption and gaus-marcov was violated."
get.assumptions(x)
}
\seealso{
\code{\link[JATSdecoder]{study.character}} for extracting multiple study characteristics at once.
}
