% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JATS_get.author.R
\name{get.author}
\alias{get.author}
\title{get.author}
\usage{
get.author(x, paste = "", short.names = FALSE, letter.convert = FALSE)
}
\arguments{
\item{x}{a NISO-JATS coded XML file or text.}

\item{paste}{if paste!="" author list is collapsed to one cell with seperator specified (e.g. paste=";").}

\item{short.names}{Logical. If TRUE fully available first names will be reduced to single letter abbreviation.}

\item{letter.convert}{Logical. If TRUE converts hexadecimal and HTML coded characters to Unicode.}
}
\value{
Character vector with the extracted author name/s.
}
\description{
Extracts author tag information from NISO-JATS coded XML file or text.
}
