% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JATS_get.contrib.R
\name{get.contrib}
\alias{get.contrib}
\title{get.contrib}
\usage{
get.contrib(x, remove.html = FALSE, letter.convert = FALSE)
}
\arguments{
\item{x}{a NISO-JATS coded XML file or text}

\item{remove.html}{Logical. If TRUE removes all html tags}

\item{letter.convert}{Logical. If TRUE converts hex and html coded characters to unicode}
}
\description{
Extract contrib tag/s from NISO-JATS coded XML file or text as vector of contributers
}
