% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JATS_get.doi.R
\name{get.doi}
\alias{get.doi}
\title{get.doi}
\usage{
get.doi(x)
}
\arguments{
\item{x}{a NISO-JATS coded XML file or text.}
}
\value{
Character string with the extracted doi.
}
\description{
Extracts articles doi from NISO-JATS coded XML file or text.
}
\seealso{
\code{\link[JATSdecoder]{JATSdecoder}} for simultaneous extraction of meta-tags, abstract, sectioned text and reference list.
}
