% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character_get.outlier.def.R
\name{get.outlier.def}
\alias{get.outlier.def}
\title{get.outlier.def}
\usage{
get.outlier.def(x)
}
\arguments{
\item{x}{text to process}
}
\description{
Extract outlier/extreme value definition/removal in standard deviations, if present in text
}
\examples{
x<-"We removed 4 extreme values that were 3 SD above mean."
get.outlier.def(x)
}
