% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character_get.R.package.R
\name{get.R.package}
\alias{get.R.package}
\title{get.R.package}
\usage{
get.R.package(x, update.package.list = FALSE)
}
\arguments{
\item{x}{text string to process.}

\item{update.package.list}{Logical. If TRUE update of list with available packages is downloaded from CRAN with utils::available.packages().}
}
\value{
Character. Vector with identified R package/s.
}
\description{
Extracts mentioned R packages from text.
}
\examples{
get.R.package("We used the R Software packages lme4 (and psych).")
}
\seealso{
\code{\link[JATSdecoder]{study.character}} for extracting multiple study characteristics at once.
}
