% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JATS_get.references.R
\name{get.references}
\alias{get.references}
\title{get.references}
\usage{
get.references(
  x,
  letter.convert = FALSE,
  remove.html = FALSE,
  extract = "full"
)
}
\arguments{
\item{x}{a NISO-JATS coded XML file or text.}

\item{letter.convert}{Logical. If TRUE converts hexadecimal and HTML coded characters to Unicode.}

\item{remove.html}{Logical. If TRUE removes all HTML tags.}

\item{extract}{part of refernces to extract (one of "full" or "title").}
}
\value{
Character vector with extracted references from reference list.
}
\description{
Extracts reference list from NISO-JATS coded XML file or text as vector of references.
}
