% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_get.sentence.with.pattern.R
\name{get.sentence.with.pattern}
\alias{get.sentence.with.pattern}
\title{get.sentence.with.pattern}
\usage{
get.sentence.with.pattern(x, patterns = c(""), tolower = TRUE)
}
\arguments{
\item{x}{sentence vector to process.}

\item{patterns}{search terms.}

\item{tolower}{Logical. If TRUE converts search terms and text to lower case.}
}
\value{
Character. Vector with sentences, that contain search pattern.
}
\description{
Returns lines with search term patterns.
}
\examples{
text<-c("This demo", "demonstrates how", "get.sentence.with.pattern works.")
get.sentence.with.pattern(text,c("Demo","example","work"))
get.sentence.with.pattern(text,c("Demo","example","work"),tolower=FALSE)
}
\keyword{internal}
