% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character_get.software.R
\name{get.software}
\alias{get.software}
\title{get.software}
\usage{
get.software(x, add.software = NULL)
}
\arguments{
\item{x}{text string to process.}

\item{add.software}{a text vector with additional software name patterns to search for.}
}
\value{
Character. Vector with identified statistical software/s.
}
\description{
Extracts mentioned software from text by dictionary search for 63 software names (object: .software_names).
}
\examples{
get.software("We used the R Software and Excel 4.0 to analyse our data.")
}
