% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character_get.stats.R
\name{get.stats}
\alias{get.stats}
\title{get.stats}
\usage{
get.stats(
  x,
  output = "both",
  stats.mode = "all",
  recalculate.p = TRUE,
  alternative = "undirected",
  estimateZ = FALSE,
  T2t = FALSE,
  R2r = FALSE,
  rm.na.col = TRUE,
  cermine = FALSE
)
}
\arguments{
\item{x}{text or JATScoded XML file to extract statistical results from}

\item{output}{Select the desired output. One of c("both","allStats","standardStats")}

\item{stats.mode}{Select subset of standardStats. One of: "all", "checkable", "computable", "uncomputable"}

\item{recalculate.p}{Logical. If TRUE recalculates p-values of standardStats if possible}

\item{alternative}{Character. Select sidedness of recomputed p-values from t-, r- and beta-values. One of c("undirected","directed","both")}

\item{estimateZ}{Logical. If TRUE detected beta-/d-value is divided by reported standard error "SE" to estimate Z-value ("Zest") for observed beta/d and recompute p-value. Note: This is only valid, if Gauss-Marcov assumptions are met and a sufficiently large sample size is used. If a Z- or t-value is detected in a report of a beta-/d-coefficient with SE, no estimation will be performed, although set to TRUE.}

\item{T2t}{Logical. If TRUE capital letter T is treated as t-statistic}

\item{R2r}{Logical. If TRUE capital letter R is treated as correlation}

\item{rm.na.col}{Logical. If TRUE removes all columns with only NA from standardStats}

\item{cermine}{Logical. If TRUE CERMINE specific letter conversion will be peformed on allStats results}
}
\description{
Extract statistical results from plain text, xml, cermxml, html, htm or docx files. The result is a list with a vector containing all identified sticked results and a matrix with containing reported standard statistics and recalculated p-values if computation is possible.
}
\examples{
x<-c("The mean difference of scale A was significant (beta=12.9, t(18)=2.5, p<.05).",
"The ANOVA yielded significant results on 
 faktor A (F(2,18)=6, p<.05, eta(g)2<-.22)",
"the correlation of x and y was r=.37.")
get.stats(x)
}
