% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JATS_get.tables.R
\name{get.tables}
\alias{get.tables}
\title{get.tables}
\usage{
get.tables(x)
}
\arguments{
\item{x}{HTML file or html text.}
}
\value{
Character vector with extracted table in html coding.
}
\description{
Extracts HTML tables as vector of tables.
}
