% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JATS_get.text.R
\name{get.text}
\alias{get.text}
\title{get.text}
\usage{
get.text(
  x,
  sectionsplit = "",
  letter.convert = TRUE,
  greek2text = FALSE,
  sentences = FALSE,
  cermine = "auto",
  rm.table = TRUE,
  rm.xref = TRUE,
  rm.media = TRUE,
  rm.graphic = TRUE,
  rm.ext_link = TRUE
)
}
\arguments{
\item{x}{a NISO-JATS coded XML file or text}

\item{sectionsplit}{search patterns for section split (forced to lower case), e.g. c("intro","method","result","discus")}

\item{letter.convert}{Logical. If TRUE converts hex and html coded characters to unicode}

\item{greek2text}{Logical. If TRUE some greek letters and special characters will be unified to textual representation. (important to extract stats)}

\item{sentences}{Logical. IF TRUE text is returned as sectioned list with sentences}

\item{cermine}{Logical. If TRUE CERMINE specific error handling and letter conversion will be applied. If set to "auto" file name ending with 'cermxml$' will set cermine=TRUE}

\item{rm.table}{Logical. If TRUE removes <table> tag from text}

\item{rm.xref}{Logical. If TRUE removes <xref> tag (citing) from text}

\item{rm.media}{Logical. If TRUE removes <media> tag from text}

\item{rm.graphic}{Logical. If TRUE removes <graphic> and <fig> tag from text}

\item{rm.ext_link}{Logical. If TRUE removes <ext link> tag from text}
}
\description{
Extract main textual content from NISO-JATS coded XML file or text as sectioned text
}
