% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JATSdecoder.R
\name{JATSdecoder}
\alias{JATSdecoder}
\title{JATSdecoder}
\usage{
JATSdecoder(
  x,
  sectionsplit = c("intro", "method", "result", "study", "experiment", "conclu",
    "implica", "discussion"),
  output = "all",
  letter.convert = TRUE,
  unify.country.name = TRUE,
  greek2text = FALSE,
  warning = FALSE
)
}
\arguments{
\item{x}{a NISO-JATS coded XML file or text}

\item{sectionsplit}{search patterns for section split (forced to lower case), e.g. c("intro","method","result","discus")}

\item{output}{selection of specific results to output c("all","title","author","affiliation","journal","volume","editor","doi","type","history","country","subject","keywords", "abstract", "sections", "text", "references")}

\item{letter.convert}{Logical. If TRUE converts hex and html coded characters to unicode}

\item{unify.country.name}{Logical. If TRUE tries to unify country name/s with list of country names from worldmap()}

\item{greek2text}{Logical. If TRUE converts and unifies several greek letters to textual representation, e.g.: alpha}

\item{warning}{Logical. If TRUE outputs a warning if processing CERMINE converted PDF files}
}
\description{
Function to extract and structure NISO-JATS coded XML file or text into a list
}
\examples{
URL <- "https://journals.plos.org/plosone/article/file?id=10.1371/journal.pone.0114876&type=manuscript"
download.file(URL,"file.xml")
JATSdecoder("file.xml")
}
