% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_ngram.R
\name{ngram}
\alias{ngram}
\title{ngram}
\usage{
ngram(x, pattern, ngram = c(-3, 3), tolower = FALSE, exact = FALSE)
}
\arguments{
\item{x}{text to process}

\item{pattern}{a search string pattern to build the ngram}

\item{ngram}{a vector of length=2 that defines the number of gram on left and right side of pattern word match}

\item{tolower}{Logical. If TRUE converts text and pattern to lower case}

\item{exact}{Logical. If TRUE only exact word matches will be proceses}
}
\description{
Extract an ngram of words around a pattern match in text
}
\examples{
text<-"One hundred twenty-eight students participated in our Study, 
that was administred in thirteen clinics."
ngram(text,pattern="study",ngram=c(-1,2))
}
