% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character_standardStats.R
\name{standardStats}
\alias{standardStats}
\title{standardStats}
\usage{
standardStats(
  x,
  stats.mode = "all",
  recalculate.p = TRUE,
  alternative = "undirected",
  estimateZ = FALSE,
  T2t = FALSE,
  rm.na.col = TRUE
)
}
\arguments{
\item{x}{result of get.stats()}

\item{stats.mode}{Select subset of standard stats. One of: "all", "checkable", "computable", "uncomputable"}

\item{recalculate.p}{Logical. If TRUE recalculates p values (for 2 sided test) if possible}

\item{alternative}{Character. Select sidedness of recomputed p-values from t-, r- and beta-values. One of c("undirected","directed","both")}

\item{estimateZ}{Logical. If TRUE detected beta-/d-value is divided by reported standard error "SE" to estimate Z-value ("Zest") for observed beta/d and recompute p-value. Note: This is only valid, if Gauss-Marcov assumptions are met and a sufficiently large sample size is used. If a Z- or t-value is detected in a report of a beta-/d-coefficient with SE, no estimation will be performed, although set to TRUE.}

\item{T2t}{Logical. If TRUE capital letter T is treated as small letter t}

\item{rm.na.col}{Logical. If TRUE removes all columns with only NA}
}
\description{
Extract and standard statistical results like Z, t, Cohen's d, F, eta^2, r, R^2, chi^2, BF_10, Q, U, H, OR, RR, beta values
}
\examples{
x<-c("t(38.8)<=>1.96, p=.002","F(2,39)<=>4, p<=>.05",
"U(2)=200, p>.25","Z<=>2.1, F(20.8,22.6)=200, p<.005, 
BF(01)<=>4","chi=3.2, r(34)<=>-.7, p<.01, R2=76\%.")
standardStats(x)
}
