% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_text2num.R
\name{text2num}
\alias{text2num}
\title{text2num}
\usage{
text2num(x)
}
\arguments{
\item{x}{text to process}
}
\description{
Convert written numbers in text to digit number (works up to 99,999)
}
\examples{
x<-"Over all one thousand two hundred twenty-eight students 
participated in our Study, that was administred in thirteen clinics."
text2num(x)
}
