% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_text2sentences.R
\name{text2sentences}
\alias{text2sentences}
\title{text2sentences}
\usage{
text2sentences(x)
}
\arguments{
\item{x}{text string to process.}
}
\value{
Character vector with sentences compiled from floating text.
}
\description{
Converts floating text to a vector with sentences via fine-tuned regular expressions.
}
\examples{
x<-"Some text with result (t(18)=1.2, p<.05). This shows how text2sentences works."
text2sentences(x)
}
