% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_vectorize.text.R
\name{vectorize.text}
\alias{vectorize.text}
\title{vectorize.text}
\usage{
vectorize.text(x)
}
\arguments{
\item{x}{text to vectorize}
}
\description{
Converts text to vector of words
}
\examples{
text<-"This demo demonstrates how vectorize.text() works."
vectorize.text(text)
}
